Timestamp fields changed: 'float32' -> 'int64'

+++ btcturk/order_book.go
@@ -4,7 +4,7 @@ import "fmt"
 
 // https://docs.btcturk.com/#order-book
 type OrderBook struct {
-	TimeStamp float64    `json:"timestamp"`
+	TimeStamp int64      `json:"timestamp"`
 	Bids      [][]string `json:"bids"`
 	Asks      [][]string `json:"asks"`
 }
